requirements.txt
Django==5.2.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9

